EDINET: FormType file
